requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.10.7
PyNaCl>=1.5.0

